Moved chat form to ilUserPrivacySettingsGUI
